Moved the provider out of the internal directory.
This move is done so that the Terraform provider (and others) can directly access the plugin.
