+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
-	"github.com/haveyoudebuggedit/terraform-provider-ovirt/internal/ovirt"
+	"github.com/haveyoudebuggedit/terraform-provider-ovirt/ovirt"
 )
 
 func main() {
