+++ ovirt/ovirt_disk.go
