+++ ovirt/ovirt_disk_attachment.go
@@ -76,7 +76,7 @@ func (p *provider) diskAttachmentResource() *schema.Resource {
 		Importer: &schema.ResourceImporter{
 			StateContext: p.diskAttachmentImport,
 		},
-		Schema:      diskAttachmentSchema,
+		Schema: diskAttachmentSchema,
 		Description: `The ovirt_disk_attachment resource attaches a single disk to a single VM. For controlling multiple attachments use ovirt_disk_attachments.
 
 ~> Do not use this resource when using ovirt_disk_attachments (plural) on the same VM as it will cause a ping-pong effect of resources being created and removed on each run.`,
