+++ ovirt/ovirt_disk_test.go
